twety three

+++ Models/ItemToWarehouse.cs
@@ -7,7 +7,7 @@ namespace Nkgjjm.Models
         [Key]
         public int Id { get; set; }
         public int WarehouseId { get; set; }
-        public int ItemId { get; set; }
+        public Int64 ItemId { get; set; }
         public double Qty { get; set; }
         public DateTime Date { get; set; }
     }
